+++ protected/humhub/modules/notification/controllers/OverviewController.php
@@ -79,7 +79,14 @@ class OverviewController extends Controller
             /* @var $notificationRecord \humhub\modules\notification\models\Notification */
 
             try {
-                $notifications[] = $notificationRecord->getBaseModel();
+                $baseModel = $notificationRecord->getBaseModel();
+
+                if($baseModel->validate()) {
+                    $notifications[] = $baseModel;
+                } else {
+                    throw new IntegrityException('Invalid base model found for notification');
+                }
+
             } catch (IntegrityException $ex) {
                 $notificationRecord->delete();
                 Yii::warning('Deleted inconsistent notification with id ' . $notificationRecord->id . '. ' . $ex->getMessage());