+++ protected/humhub/modules/space/messages/de/notification.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  'You were added to Space {spaceName}' => 'Du wurden zum Space {spaceName} hinzugefügt',
+  'You were added to Space {spaceName}' => 'Du wurdest zum Space {spaceName} hinzugefügt',
   '{displayName} accepted your invite for the space {spaceName}' => '{displayName} hat deine Einladung für den {spaceName} akzeptiert',
   '{displayName} approved your membership for the space {spaceName}' => '{displayName} hat deine Mitgliedschaft für den {spaceName} bestätigt',
   '{displayName} changed your role to {roleName} in the space {spaceName}.' => '{displayName} hat im Space {spaceName} Deine Rolle auf {roleName} geändert.