+++ protected/humhub/docs/CHANGELOG.md
@@ -9,6 +9,7 @@ HumHub Change Log
 - Enh: Improved theme parent lookup performance
 - Enh: Added auto file attachment in `humhub\modules\content\widgets\richtext\ProsemirrorRichTextProcessor`
 - Fix: z-index issue with fixed richtext menu
+- Fix #3294: space picker encoding
 
 
 1.3.2  (September 4, 2018)